+++ fs/ceph/dir.c
@@ -323,7 +323,8 @@ more:
 
 		dout("readdir fetching %llx.%llx frag %x offset '%s'\n",
 			(unsigned long long) ceph_ino(inode),
-			ceph_snap(inode), frag, fi->last_name);
+			(unsigned long long) ceph_snap(inode),
+			frag, fi->last_name);
 		req = ceph_mdsc_create_request(mdsc, op, USE_AUTH_MDS);
 		if (IS_ERR(req))
 			return PTR_ERR(req);
