+++ fs/ceph/file.c
@@ -137,7 +137,8 @@ int ceph_open(struct inode *inode, struct file *file)
 
 	dout("open inode %p ino %llx.%llx file %p flags %d (%d)\n", inode,
 		(unsigned long long) ceph_ino(inode),
-		ceph_snap(inode), file, flags, file->f_flags);
+		(unsigned long long) ceph_snap(inode),
+		file, flags, file->f_flags);
 	fmode = ceph_flags_to_mode(flags);
 	wanted = ceph_caps_for_mode(fmode);
 
@@ -203,7 +204,7 @@ int ceph_open(struct inode *inode, struct file *file)
 	ceph_mdsc_put_request(req);
 	dout("open result=%d on %llx.%llx\n", err,
 		(unsigned long long) ceph_ino(inode),
-		ceph_snap(inode));
+		(unsigned long long) ceph_snap(inode));
 out:
 	return err;
 }
@@ -641,7 +642,8 @@ static ssize_t ceph_aio_read(struct kiocb *iocb, const struct iovec *iov,
 
 	dout("aio_read %p %llx.%llx %llu~%u trying to get caps on %p\n", inode,
 		(unsigned long long) ceph_ino(inode),
-		ceph_snap(inode), pos, (unsigned) len, inode);
+		(unsigned long long) ceph_snap(inode),
+		pos, (unsigned) len, inode);
 again:
 	__ceph_do_pending_vmtruncate(inode);
 	if (fi->fmode & CEPH_FILE_MODE_LAZY)
@@ -653,7 +655,8 @@ again:
 		goto out;
 	dout("aio_read %p %llx.%llx %llu~%u got cap refs on %s\n", inode,
 		(unsigned long long) ceph_ino(inode),
-		ceph_snap(inode), pos, (unsigned)len, ceph_cap_string(got));
+		(unsigned long long) ceph_snap(inode),
+		pos, (unsigned)len, ceph_cap_string(got));
 
 	if ((got & (CEPH_CAP_FILE_CACHE|CEPH_CAP_FILE_LAZYIO)) == 0 ||
 	    (iocb->ki_filp->f_flags & O_DIRECT) ||
@@ -667,7 +670,8 @@ again:
 out:
 	dout("aio_read %p %llx.%llx dropping cap refs on %s = %d\n", inode,
 		(unsigned long long) ceph_ino(inode),
-		ceph_snap(inode), ceph_cap_string(got), (int) ret);
+		(unsigned long long) ceph_snap(inode),
+		ceph_cap_string(got), (int) ret);
 	ceph_put_cap_refs(ci, got);
 
 	if (checkeof && ret >= 0) {
@@ -721,7 +725,8 @@ retry_snap:
 	__ceph_do_pending_vmtruncate(inode);
 	dout("aio_write %p %llx.%llx %llu~%u getting caps. i_size %llu\n",
 		inode, (unsigned long long) ceph_ino(inode),
-		ceph_snap(inode), pos, (unsigned) iov->iov_len, inode->i_size);
+		(unsigned long long) ceph_snap(inode), pos,
+		(unsigned) iov->iov_len, inode->i_size);
 	if (fi->fmode & CEPH_FILE_MODE_LAZY)
 		want = CEPH_CAP_FILE_BUFFER | CEPH_CAP_FILE_LAZYIO;
 	else
@@ -732,8 +737,8 @@ retry_snap:
 
 	dout("aio_write %p %llx.%llx %llu~%u  got cap refs on %s\n", inode,
 		(unsigned long long) ceph_ino(inode),
-		ceph_snap(inode), pos, (unsigned) iov->iov_len,
-		ceph_cap_string(got));
+		(unsigned long long) ceph_snap(inode),
+		pos, (unsigned) iov->iov_len, ceph_cap_string(got));
 
 	if ((got & (CEPH_CAP_FILE_BUFFER|CEPH_CAP_FILE_LAZYIO)) == 0 ||
 	    (iocb->ki_filp->f_flags & O_DIRECT) ||
@@ -779,15 +784,16 @@ retry_snap:
 out_put:
 	dout("aio_write %p %llx.%llx %llu~%u  dropping cap refs on %s\n",
 	     inode, (unsigned long long) ceph_ino(inode),
-	     ceph_snap(inode), pos, (unsigned) iov->iov_len,
-	     ceph_cap_string(got));
+	     (unsigned long long) ceph_snap(inode), pos,
+	     (unsigned) iov->iov_len, ceph_cap_string(got));
 	ceph_put_cap_refs(ci, got);
 
 out:
 	if (ret == -EOLDSNAPC) {
 		dout("aio_write %p %llx.%llx %llu~%u got EOLDSNAPC, retrying\n",
 		     inode, (unsigned long long) ceph_ino(inode),
-		     ceph_snap(inode), pos, (unsigned) iov->iov_len);
+		     (unsigned long long) ceph_snap(inode), pos,
+		     (unsigned) iov->iov_len);
 		goto retry_snap;
 	}
