+++ net/ceph/osd_client.c
@@ -1898,7 +1898,8 @@ int ceph_osdc_readpages(struct ceph_osd_client *osdc,
 	int rc = 0;
 
 	dout("readpages on ino %llx.%llx on %llu~%llu\n",
-		(unsigned long long) vino.ino, vino.snap, off, *plen);
+		(unsigned long long) vino.ino,
+		(unsigned long long) vino.snap, off, *plen);
 	req = ceph_osdc_new_request(osdc, layout, vino, off, plen,
 				    CEPH_OSD_OP_READ, CEPH_OSD_FLAG_READ,
 				    NULL, 0, truncate_seq, truncate_size, NULL,
